More bug fixed
It can happen that the collisions penetrate too mucho so after it, it can make the sistem collide again the next instant. What I did was to delete to the position on that time the velocity of the objects. More realistic would be to delete the distance to the collision border after the first collision.